test: improve integration tests with different containers
